refactor organizer dashboard
