+++ components/dashboard/organizer/tracks/TracksTab.tsx
@@ -1,7 +1,6 @@
 "use client";
 
 import { useState, useEffect } from "react";
-import { Card, CardContent, CardHeader, CardTitle } from "@/components/ui/card";
 import { ScoringConfigForm } from "@/components/scoring/scoring-config-form";
 import {
   EventDetails,
@@ -159,27 +158,19 @@ export function TracksTab({
 
   return (
     <div className="space-y-6">
-      {/* Track Configuration with integrated judging modes */}
-      <Card>
-        <CardHeader>
-          <CardTitle>Event Tracks</CardTitle>
-        </CardHeader>
-        <CardContent>
-          <p className="text-sm text-muted-foreground mb-6">
-            Configure your event tracks below. For each track, you can set up
-            scoring criteria and choose the judging interface (traditional
-            scoring or investment decision).
-          </p>
+      <p className="text-sm text-muted-foreground mb-6">
+        Configure your event tracks below. For each track, you can set up
+        scoring criteria and choose the judging interface (traditional
+        scoring or investment decision).
+      </p>
 
-          <ScoringConfigForm
-            initialConfig={scoringConfig}
-            isSubmitting={saving}
-            onSave={handleScoreUpdate}
-            tracks={tracks}
-            onJudgingModeUpdate={handleJudgingModeUpdate}
-          />
-        </CardContent>
-      </Card>
+      <ScoringConfigForm
+        initialConfig={scoringConfig}
+        isSubmitting={saving}
+        onSave={handleScoreUpdate}
+        tracks={tracks}
+        onJudgingModeUpdate={handleJudgingModeUpdate}
+      />
     </div>
   );
 }
